Punkty 13-14: refraktoryzacja main, ukrycie klasy Node
